Fixed bugs - READ
fixed the responsiveness of the site + made all inputs locked at the beginning (readonly is true). After clicking on a level of difficulty, the inputs are assigned their respective values and their readonly status changes: if the cell should be a number other than 0, it's assigned that value and locked for editing. Otherwise, it's value is null and it's unlocked for editing. Next step is to handle user input and create the logic for the solver.